requirements.txt
pinecone-client==5.0.1
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
sentence-transformers==2.7.0
huggingface-hub==0.23.0
openai==1.3.7
numpy==1.24.3
pandas==2.1.4
pytest==7.4.3
requests==2.31.0
typing-extensions==4.8.0
qdrant-client==1.7.0
# PDF processing dependencies
PyPDF2==3.0.1
tiktoken==0.5.1 